+++ bot/main.go
@@ -1,20 +1,20 @@
 package bot
 
 import (
-	"bookq.xyz/mercari-watchdog/utils"
+	"bookq.xyz/mercari-watchdog/tools"
 	Pichubot "github.com/0ojixueseno0/go-Pichubot"
 )
 
-var OperationChan chan utils.PushMsg
-var Push1to2Chan chan utils.PushMsg
-var Push3to4Chan chan utils.PushMsg
-var Push5upChan chan utils.PushMsg
+var OperationChan chan tools.PushMsg
+var Push1to2Chan chan tools.PushMsg
+var Push3to4Chan chan tools.PushMsg
+var Push5upChan chan tools.PushMsg
 
 func Boot() {
-	OperationChan = make(chan utils.PushMsg, 10)
-	Push1to2Chan = make(chan utils.PushMsg, 10)
-	Push3to4Chan = make(chan utils.PushMsg, 10)
-	Push5upChan = make(chan utils.PushMsg, 10)
+	OperationChan = make(chan tools.PushMsg, 10)
+	Push1to2Chan = make(chan tools.PushMsg, 10)
+	Push3to4Chan = make(chan tools.PushMsg, 10)
+	Push5upChan = make(chan tools.PushMsg, 10)
 
 	Pichubot.Listeners.OnGroupMsg = append(Pichubot.Listeners.OnGroupMsg, handlerHelp, handlerGroupMsg)
 	Pichubot.Listeners.OnGroupRequest = append(Pichubot.Listeners.OnGroupRequest, handlerGroupRequest)
