+++ datatype/analysistask/definition.go
