+++ tools/filters.go
@@ -1,6 +1,7 @@
 package tools
 
 import (
+	"reflect"
 	"strings"
 
 	"github.com/bookqaq/goForMercari/mercarigo"
@@ -34,4 +35,18 @@ func PriceFilter(price [2]int, data []mercarigo.MercariItem) []mercarigo.Mercari
 		return data
 	}
 	return result
+}
+
+// Delete item in ordered array src that return true in reflect.DeepEqual(item, value), return lenght deleted
+func DeleteInvalidItem[T any](src []T, value T) int {
+	deleted, formerpt, length := 0, 0, len(src)
+	for i := 0; i < length; i++ {
+		if reflect.DeepEqual(src[i], value) {
+			deleted++
+		} else {
+			src[formerpt] = src[i]
+			formerpt++
+		}
+	}
+	return deleted
 }
