+++ tools/main.go
@@ -1,0 +1,6 @@
+package tools
+
+type PushMsg struct {
+	Dst int64
+	S   []string
+}
